# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create a warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /warehouses/{id}:
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update a warehouse
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Deactivate a warehouse
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Warehouse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, isDefault, isActive, createdAt]
    WarehouseInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        isDefault: { type: boolean }
        isActive: { type: boolean }
...